Data Module

HVPobs.Data.read_hvp_data — Function

@doc raw readhvpdata(path::String, id::String)

This function reads the HVP data used in the g-2 analysis. It takes as input a path to the data and en ensemble id and it returns a CData structure.

@example cdata = read_hvp_data(path, "H101")`

HVPobs.Data.read_mesons_data — Function

@doc raw readmesondata(path::String, id::String)

This function reads the 2-point function data in the format used for spectrum computations. It takes as input a path to the data and en ensemble id and it returns a CData structure.

@example cdata = read_mesons_data(path, "H101")`

HVPobs.Data.read_ms — Function

read_ms(path::String; id::Union{String, Nothing}=nothing, dtr::Int64=1, obs::String="Y")

Reads openQCD ms dat files at a given path. This method return YData:

t(t): flow time values

obs(icfg, x0, t): the time-slice sums of the densities of the observable (Wsl, Ysl or Qsl)

vtr: vector that contains trajectory number

id: ensemble id

dtr = dtr_cnfg / dtr_ms, where dtr_cnfg is the number of trajectories computed before saving the configuration. dtr_ms is the same but applied to the ms.dat file.

Examples:

Y = read_ms(path)

HVPobs.Data.read_ms1 — Function

read_ms1(path::String; v::String="1.2")

Reads openQCD ms1 dat files at a given path. This method returns a matrix W[irw, icfg] that contains the reweighting factors, where irw is the rwf index and icfg the configuration number. The function is compatible with the output files of openQCD v=1.2, 1.4 and 1.6. Version can be specified as argument.

Examples:

read_ms1(path)
read_ms1(path, v="1.4")
read_ms1(path, v="1.6")
read_ms1(path, v="2.0")