Obs Module

ObsPrimary

HVPobs.Obs.Corr — Type

Corr(a::Vector{uwreal}, id::String, gamma::String)

The struct Corr stores data, ensemble id and gamma structure for two and three-point correlation functions.

HVPobs.Obs.corr_obs — Function

corr_obs(cd::CData; real::Bool=true, rw::Union{Array{Float64,2}, Vector{Array{Float64,2}}, Nothing}=nothing, L::Int64=1, nms::Union{Int64, Nothing}=nothing)

Creates a Corr struct with the given CData struct cdata (read from read_hvp_data or read_mesons_data) for a single or multiple replicas. The flag real select the real or imaginary part of the correlator. If rw is specified, the method applies reweighting. rw is passed as a matrix of Float64 (read_ms1) The correlator can be normalized with the volume factor if L is fixed. The flag nms can be used to specify the maximum number of existing configurations in a given ensemble.

data = read_hvp_data(path, id)
rw = read_ms1(path_rw)

corr = corr_obs(data)
corr_r = corr_obs(data, rw=rw)

HVPobs.Obs.comp_t0 — Function

comp_t0(Y::YData, plat::Vector{Int64}; L::Int64, pl::Bool=false, rw::Union{Matrix{Float64}, Nothing}=nothing, npol::Int64=2, wpm::Union{Dict{Int64,Vector{Float64}},Dict{String,Vector{Float64}}, Nothing}=nothing, info::Bool=false)

comp_t0(Y::Vector{YData}, plat::Vector{Int64}; L::Int64, pl::Bool=false, rw::Union{Vector{Matrix{Float64}}, Nothing}=nothing, npol::Int64=2, wpm::Union{Dict{Int64,Vector{Float64}},Dict{String,Vector{Float64}}, Nothing}=nothing, info::Bool=false)

Computes t0 using the energy density of the action Ysl(Yang-Mills action). t0 is computed in the plateau plat. A polynomial interpolation in t is performed to find t0, where npol is the degree of the polynomial (linear fit by default)

The flag pl allows to show the plot.

The flag info provides extra output that contains information about the primary observables. The function returns the primary observables $<WY>$ and $<W>$ (it returns the observable <Y> if rw=nothing)

#Single replica
Y = read_ms(path)
rw = read_ms(path_rw)

t0, Yobs = comp_t0(Y, [38, 58], L=32, info=true)
t0_r, WYobs, Wobs = comp_t0(Y, [38, 58], L=32, rw=rw, info=true)

#Two replicas
Y1 = read_ms(path1)
Y2 = read_ms(path2)
rw1 = read_ms(path_rw1)
rw2 = read_ms(path_rw2)

t0 = comp_t0([Y1, Y2], [38, 58], L=32, pl=true)
t0_r = comp_t0(Y, [38, 58], L=32, rw=[rw1, rw2], pl=true)

HVPobs.Obs.comp_fvc — Function

comp_fvc(path::String)

Computes the finite volume corrections from the jackknife samples and returns the uwreal object as an Array{uwreal}(n2max, Thalf), where n2max is the number (squared) of pion wrapping around the torus and Thalf is half of the lattice temporal extent. The data file to be passed are typically named fvcorrblatgslrng0000err3ensIDwin0.dat

path = /path/to/fv_corr_blat_gslrng0000_err3_ensID.dat
data = comp_fvc(path)

ObsSpectrum

HVPobs.Obs.meff — Function

meff(obs::Vector{uwreal}, plat::Vector{Int64}; pl::Bool=false, data::Bool=false, wpm::Union{Dict{Int64,Vector{Float64}},Dict{String,Vector{Float64}}, Nothing}=nothing)
meff(corr::Corr, plat::Vector{Int64}; pl::Bool=true, data::Bool=false, wpm::Union{Dict{Int64,Vector{Float64}},Dict{String,Vector{Float64}}, Nothing}=nothing)

Computes effective mass for a given correlator corr at a given plateau plat. Correlator can be passed as an Corr struct or Vector{uwreal}.

The flags pl and data allow to show the plots and return data as an extra result.

data = read_mesons(path, "G5", "G5")
corr_pp = corr_obs.(data)
m = meff(corr_pp[1], [50, 60], pl=false)

ObsTools

HVPobs.Obs.plat_av — Function

@doc raw plat_av(obs::Vector{uwreal}, plat::Vector{Int64}; wpm::Union{Dict{Int64,Vector{Float64}},Dict{String,Vector{Float64}}, Nothing}=nothing)

This function performs the plateau average of obs over the plateau plat. The wpm flag can be used to control error analysis according to the ADerrors documentation.

HVPobs.Obs.frwd_bckwrd_symm! — Function

@doc raw frwdbckwrdsymm!(obs::Vector{uwreal}) frwdbckwrdsymm!(corr::Corr)

This function performs the forward-backward symmetrization of a given obs passed as input.

ObsImprovement

HVPobs.Obs.improve_corr_vkvk! — Function

improve_corr_vkvk!(vkvk::Vector{uwreal}, vkt0k::Vector{uwreal}, cv::Union{Float64,uwreal}; std::Bool=false)
improve_corr_vkvk!(vkvk::Corr, t0tk::Corr, cv::Union{Float64,uwreal}; std::Bool=false)

Given the vector-vector vkvkand vector-tensor vkt0k correlators as input, together with the cv improvement coefficient, this function overwrite the vkvk correlator with the Symanzik improved version. The flag std controls the derivative discretization of the vkt0k correlators. If false, it uses the improved version of the derivative, if true it uses the standard symmetric derivative.

```@example
cdata_vv = read_hvp_data(path_to_vv, id)
cdata_vt = read_hvp_data(path_to_vt, id)

rw = read_ms1(path, v="1.4")

corr_vv = corr_obs(cdata_vv, rw=rw)
corr_vt = corr_obs(cdata_vt, rw=rw)

cv = cv_loc(beta)
improve_corr_vkvk!(corr_vv, corr_vt, cv )
```

HVPobs.Obs.improve_corr_vkvk_cons! — Function

improve_corr_vkvk_cons!(vkvk::Vector{uwreal}, vkt0k_l::Vector{uwreal}, vkt0k_c::Vector{uwreal}, cv_l::Union{Float64,uwreal}, cv_c::Union{Float64,uwreal}; std::Bool=false)
improve_corr_vkvk_cons!(vkvk::Corr, t0tk_l::Corr, t0tk_c::Corr, cv_l::Union{Float64,uwreal}, cv_c::Union{Float64,uwreal}; std::Bool=false)

Given the local-conserved vector-vector vkvk, the local-conserved vector-tensor vkt0k and the local-local vector-tensor vkt0k correlators as input, together with the cvlocal and cvcons improvement coefficient, this function overwrite the local-conserved vkvk correlator with the Symanzik improved version. The flag std controls the derivative discretization of the vkt0k correlators. If false, it uses the improved version of the derivative, if true it uses the standard symmetric derivative.

```@example
cdata_vv = read_hvp_data(path_to_vv, id)
cdata_vt_loc = read_hvp_data(path_to_vt_loc, id)
cdata_vt_cons = read_hvp_data(path_to_vt_cons, id)

rw = read_ms1(path, v="1.4")

corr_vv = corr_obs(cdata_vv, rw=rw)
corr_vt_loc = corr_obs(cdata_vt_loc, rw=rw)
corr_vt_cons = corr_obs(cdata_vt_cons, rw=rw)

cv_loc  = cv_loc(beta)
cv_cons = cv_cons(beta)

improve_corr_vkvk!(corr_vv, corr_vt_loc, corr_vt_cons, cv_loc, cv_cons)
```

HVPobs.Obs.ZV — Function

ZV(beta::Float64)

Given the coupling beta, this function returns the ZV renormalization constant for the vector current based on the results of 1811.08209.

HVPobs.Obs.ZV_set2 — Function

ZV_set2(beta::Float64)

Given the coupling beta, this function returns the ZV renormalization constant for the vector current based on the SF results of 2010.09539.

HVPobs.Obs.cv_loc — Function

cv_loc(beta::Float64)

Given the coupling beta, this function returns the cv improvement coefficient for the local vector current based on the results of 1811.08209.

HVPobs.Obs.cv_loc_set2 — Function

cvlocset2(beta::Float64)

Given the coupling beta, this function returns the cv improvement coefficient for the local vector current based on the SF results of 2010.09539.

HVPobs.Obs.cv_cons — Function

cv_cons(beta::Float64)

Given the coupling beta, this function returns the cv improvement coefficient for the conserved vector current based on the results of 1811.08209.

HVPobs.Obs.cv_cons_set2 — Function

cvconsset2(beta::Float64)

Given the coupling beta, this function returns the cv improvement coefficient for the conserved vector current based on the SF results of 2010.09539.

HVPobs.Obs.bv — Function

bv(beta::Float64)

Given the coupling beta, this function returns the bv improvement coefficient for the vector carrent based on the results of 1811.08209.

HVPobs.Obs.bv_set2 — Function

bv_set2(beta::Float64)

Given the coupling beta, this function returns the bv coefficient for the improvement of the ZV renormalization constants. Values from 1805.07401

HVPobs.Obs.bv_bar — Function

bv_bar(beta::Float64)

Given the coupling beta, this function returns the bv_bar improvement coefficient for the vector carrent based on the results of 1811.08209.

HVPobs.Obs.bv_bar_set2 — Function

bvbarset2(beta::Float64)

Given the coupling beta, this function returns the bv_bar coefficient for the improvement of the Zv renormalization constants. Values from 1805.07401